Add QueueClassic specs
